refactor(gateway): Q11.d — extract rest_v1/tasks/labels.rs (GAR-635)
Move task-label CRUD section (plan 0078 / GAR-536, ~524 LOC) from
rest_v1/tasks/mod.rs into rest_v1/tasks/labels.rs.

- New: tasks/labels.rs (546 LOC) — 5 handlers + 6 DTOs + is_valid_hex_color
- mod.rs: 2857 → 2341 lines; add `pub mod labels; pub use labels::{...}`
- openapi.rs: update 5 handler paths to super::tasks::labels::*

Zero behavior change — pure structural refactor (plan 0138).
