+++ crates/garraia-gateway/src/rest_v1/openapi.rs
@@ -132,11 +132,11 @@ impl Modify for SecurityAddon {
         super::tasks::assignees::add_task_assignee,
         super::tasks::assignees::list_task_assignees,
         super::tasks::assignees::remove_task_assignee,
-        super::tasks::create_task_label,
-        super::tasks::list_task_labels,
-        super::tasks::delete_task_label,
-        super::tasks::assign_task_label,
-        super::tasks::remove_task_label_from_task,
+        super::tasks::labels::create_task_label,
+        super::tasks::labels::list_task_labels,
+        super::tasks::labels::delete_task_label,
+        super::tasks::labels::assign_task_label,
+        super::tasks::labels::remove_task_label_from_task,
         super::tasks::subscribe_to_task,
         super::tasks::list_task_subscriptions,
         super::tasks::unsubscribe_from_task,
